Acmp 929 task. Minor fixes

+++ src/java/dev/avyguzov/acmp/Task929.java
@@ -5,10 +5,10 @@ import java.util.Scanner;
 // https://acmp.ru/index.asp?main=task&id_task=929
 public class Task929 {
 
-    public static int solutionMinScore(int score) {
+    public static int solutionMinScore(long score) {
         int result = 0;
 
-        int cubeSidesCount;
+        long cubeSidesCount;
         int i = 6;
         while (i > 0) {
             cubeSidesCount = score / i;
@@ -27,13 +27,13 @@ public class Task929 {
         return result;
     }
 
-    public static int solutionMaxScore(int score) {
+    public static long solutionMaxScore(long score) {
         return score * 6;
     }
 
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
-        int score = s.nextInt();
+        long score = s.nextLong();
 
         System.out.println(Task929.solutionMinScore(score) + " " + Task929.solutionMaxScore(score));
         s.close();
